[DROOLS-3341] Tests improved.

+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-client/src/main/java/org/drools/workbench/screens/scenariosimulation/client/models/ScenarioGridModel.java
@@ -498,7 +498,7 @@ public class ScenarioGridModel extends BaseGridData {
      * @param className
      * @return
      */
-    public long getInstancesCount(String className) {
+    public int getInstancesCount(String className) {
         return simulation.getSimulationDescriptor().getUnmodifiableFactMappings()
                 .stream()
                 .filter(factMapping-> factMapping.getFactIdentifier().getClassName().equals(className))
